ci (GitHub Actions): Add new build and test workflow
Introduce a new GitHub Actions workflow named "Build and Test main" in the `dotnet.yml` file. The workflow is triggered on pushes to the `main` branch, excluding changes to markdown, gitignore, and gitattributes files. It also allows manual triggering with a required input for the reason for running the workflow.

The workflow defines a single job, `build-and-test`, which runs on `ubuntu-latest` and includes steps to:
- Check out the repository.
- Fetch all history for tags and branches.
- Print the reason for a manual run if applicable.
- Set up the .NET Core SDK using a specified global.json file.
- Restore workloads and dependencies.
- Build the solution in Release configuration.
- Run tests with specified settings.
- Install and use GitVersion to manage versioning.
- Print the full semantic version.
- Use Codecov to upload code coverage results.
- Upload test results as an artifact, ensuring this step runs even if there are test failures.