ci (GitHub Actions): Add "Build and Test dev" workflow
A new GitHub Actions workflow named "Build and Test dev" has been added.
The workflow triggers on pushes to the `dev` branch and ignores changes to markdown, gitignore, and gitattributes files. It can also be manually triggered with a reason input.

The workflow defines a single job `build-and-test` that runs on `ubuntu-latest` and sets several environment variables to configure the .NET build environment. The job consists of multiple steps:
- Checkout the repository.
- Fetch all history for all tags and branches.
- Print the reason for a manual run if triggered manually.
- Setup the .NET Core SDK using a global.json file.
- Restore .NET workloads.
- Restore project dependencies.
- Build the project in Release configuration without restoring dependencies.
- Run tests without building the project again, using a specified runsettings.xml file.
- Install and use GitVersion to manage versioning.
- Print the full semantic version generated by GitVersion.
- Use Codecov to upload code coverage results.
- Upload test results as an artifact, ensuring this step runs even if previous steps fail.