ci (GitHub Actions): Add code formatting workflow
Introduce a new GitHub Actions workflow configuration in `formatting.yml`.
This workflow, named "Code formatting", triggers on pushes to `main` and `dev`
branches for changes in `.cs`, `.js`, and `.css` files, as well as C# test files.

The workflow includes a `code-formatting` job that runs on `ubuntu-latest` and
is conditioned to run only if the branch is not `main` and the repository is not
`mpaulosky/AspireBlogApp`. The job performs the following steps:

- Checkout the repository using `actions/checkout@v4` with a `fetch-depth` of 0.
- Retrieve the list of changed files compared to the `main` branch.
- Filter the list to include only changed `.cs` files.
- Append specific formatting rules to `.editorconfig` if there are changed C# files.
- Setup the .NET environment using `actions/setup-dotnet@v4` with version `8.0.x`.
- Run `dotnet format` on the solution file, excluding migrations.
- Filter the list to include only changed `.js` and `.css` files.
- Setup the Node.js environment using `actions/setup-node@v4` with the latest LTS version.
- Install and run `prettier` to format the JS/CSS files.
- Attempt to commit and push the formatting changes, continuing on error if no files were formatted.