ci (GitHub Actions): Add CodeQL analysis workflow
Added a new GitHub Actions workflow configuration file named `codeql-analysis.yml` to set up CodeQL analysis for the repository. The workflow is triggered on pushes and pull requests to the `main` and `develop` branches for `.cs` and `.csproj` files. A manual trigger (`workflow_dispatch`) is also included, allowing the workflow to be run manually with a specified reason. Scheduled runs are configured to occur every Friday at 00:31 UTC.

The workflow defines a single job named `analyze` that runs on the latest Ubuntu environment. Permissions are set to allow reading actions and contents, and writing security events. The job strategy is set to not fail fast and includes a matrix to specify the language as `csharp`.

The steps in the job include:
- Checking out the repository using `actions/checkout@v4`.
- Initializing CodeQL using `github/codeql-action/init@v3` with the specified language.
- Running the autobuild process using `github/codeql-action/autobuild@v3`.
- Performing the CodeQL analysis using `github/codeql-action/analyze@v3`.